requirements.txt
streamlit
streamlit-lottie
tensorflow
numpy
Pillow
requests